+++ RestAPI/Servicos/IPessoaService.cs
@@ -5,9 +5,9 @@ namespace RestAPI.Servicos
     public interface IPessoaService
     {
         Pessoa CriarPessoa(Pessoa pessoa);
-        Pessoa ProcurarPorId(long id);
+        Pessoa ProcurarPorId(int id);
         List<Pessoa> ListarTodasPessoas();
         Pessoa AtualizarPessoa(Pessoa pessoa);
-        void DeletarPessoa(long id);
+        void DeletarPessoa(int id);
     }
 }
